+++ alacritty/src/display/mod.rs
@@ -6,14 +6,14 @@ use std::fmt::{self, Formatter};
 use std::mem::{self, ManuallyDrop};
 use std::num::NonZeroU32;
 use std::ops::{Deref, DerefMut};
-#[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
 use std::sync::atomic::Ordering;
+use std::time::{Duration, Instant};
 
 use glutin::context::{NotCurrentContext, PossiblyCurrentContext};
 use glutin::prelude::*;
 use glutin::surface::{Rect as DamageRect, Surface, SwapInterval, WindowSurface};
 
-use log::{debug, info, warn};
+use log::{debug, info};
 use parking_lot::MutexGuard;
 use serde::{Deserialize, Serialize};
 use winit::dpi::PhysicalSize;
@@ -46,10 +46,11 @@ use crate::display::damage::RenderDamageIterator;
 use crate::display::hint::{HintMatch, HintState};
 use crate::display::meter::Meter;
 use crate::display::window::Window;
-use crate::event::{Mouse, SearchState};
+use crate::event::{Event, EventType, Mouse, SearchState};
 use crate::message_bar::{MessageBuffer, MessageType};
 use crate::renderer::rects::{RenderLine, RenderLines, RenderRect};
 use crate::renderer::{self, GlyphCache, Renderer};
+use crate::scheduler::{Scheduler, TimerId, Topic};
 use crate::string::{ShortenDirection, StrShortener};
 
 pub mod content;
@@ -367,6 +368,9 @@ pub struct Display {
     /// The ime on the given display.
     pub ime: Ime,
 
+    /// The state of the timer based for frame scheduling.
+    pub frame_timer_info: FrameTimerInfo,
+
     // Mouse point position when highlighting hints.
     hint_mouse_point: Option<Point>,
 
@@ -487,14 +491,8 @@ impl Display {
             (Vec::new(), Vec::new())
         };
 
-        // We use vsync everywhere except wayland.
-        if !is_wayland {
-            if let Err(err) =
-                surface.set_swap_interval(&context, SwapInterval::Wait(NonZeroU32::new(1).unwrap()))
-            {
-                warn!("Error setting vsync: {:?}", err);
-            }
-        }
+        // Disable vsync.
+        let _ = surface.set_swap_interval(&context, SwapInterval::DontWait);
 
         Ok(Self {
             window,
@@ -510,6 +508,7 @@ impl Display {
             vi_highlighted_hint: None,
             is_wayland,
             cursor_hidden: false,
+            frame_timer_info: Default::default(),
             visual_bell: VisualBell::from(&config.bell),
             colors: List::from(&config.colors),
             pending_update: Default::default(),
@@ -751,6 +750,7 @@ impl Display {
     pub fn draw<T: EventListener>(
         &mut self,
         mut terminal: MutexGuard<'_, Term<T>>,
+        scheduler: &mut Scheduler,
         message_buffer: &MessageBuffer,
         config: &UiConfig,
         search_state: &SearchState,
@@ -969,7 +969,9 @@ impl Display {
         // Frame event should be requested before swaping buffers, since it requires surface
         // `commit`, which is done by swap buffers under the hood.
         #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        self.request_frame(&self.window);
+        if self.is_wayland {
+            self.request_frame(scheduler);
+        }
 
         // Clearing debug highlights from the previous frame requires full redraw.
         self.swap_buffers();
@@ -980,6 +982,10 @@ impl Display {
             // will block to synchronize (this is `glClear` in Alacritty), which causes a
             // permanent one frame delay.
             self.renderer.finish();
+        }
+
+        if !self.is_wayland {
+            self.request_frame(scheduler);
         }
 
         self.damage_rects.clear();
@@ -1376,23 +1382,38 @@ impl Display {
     }
 
     /// Requst a new frame for a window on Wayland.
-    #[inline]
-    #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-    fn request_frame(&self, window: &Window) {
-        let surface = match window.wayland_surface() {
-            Some(surface) => surface,
-            None => return,
-        };
+    fn request_frame(&mut self, scheduler: &mut Scheduler) {
+        // Mark that we've used a frame.
+        self.window.has_frame.store(false, Ordering::Relaxed);
 
-        let should_draw = self.window.should_draw.clone();
+        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
+        if let Some(surface) = self.window.wayland_surface() {
+            let has_frame = self.window.has_frame.clone();
+            // Request a new frame.
+            surface.frame().quick_assign(move |_, _, _| {
+                has_frame.store(true, Ordering::Relaxed);
+            });
 
-        // Mark that window was drawn.
-        should_draw.store(false, Ordering::Relaxed);
+            return;
+        }
 
-        // Request a new frame.
-        surface.frame().quick_assign(move |_, _, _| {
-            should_draw.store(true, Ordering::Relaxed);
-        });
+        // Get the display vblank interval in micro seconds.
+        let monitor_vblank_interval = Duration::from_micros(
+            (1e6 / (self
+                .window
+                .current_monitor()
+                .and_then(|monitor| monitor.refresh_rate_millihertz())
+                .unwrap_or(60_000) as f64
+                / 1e3)) as u64,
+        );
+
+        let swap_delay = self.frame_timer_info.compute_delay(monitor_vblank_interval);
+
+        let window_id = self.window.id();
+        let timer_id = TimerId::new(Topic::Frame, window_id);
+        let event = Event::new(EventType::Frame, window_id);
+
+        scheduler.schedule(event, swap_delay, false, timer_id);
     }
 }
 
@@ -1531,6 +1552,54 @@ impl<T> Deref for Replaceable<T> {
 impl<T> DerefMut for Replaceable<T> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         self.get_mut()
+    }
+}
+
+/// The frame timer state.
+#[derive(Default)]
+pub struct FrameTimerInfo {
+    /// The optimal delay for the renderer.
+    swap_delay: Option<Duration>,
+
+    /// The time stamp of the previous contiguos frame.
+    frame_timestamp: Option<Instant>,
+
+    /// The refresh rate we've used to compute the `swap_delay`.
+    refresh_interval: Duration,
+}
+
+impl FrameTimerInfo {
+    /// Compute the delay that we should use to achieve the target frame
+    /// rate.
+    pub fn compute_delay(&mut self, refresh_interval: Duration) -> Duration {
+        // If the refresh rate has changed, reset the timings.
+        if self.refresh_interval != refresh_interval || self.swap_delay.is_none() {
+            self.refresh_interval = refresh_interval;
+            self.frame_timestamp = Some(Instant::now());
+            self.swap_delay = Some(refresh_interval);
+            return refresh_interval;
+        }
+
+        // At this point we know that `swap_delay` has some value.
+        let swap_delay = self.swap_delay.unwrap();
+
+        let new_swap_delay = match self.frame_timestamp {
+            Some(last_frame_timestamp) => {
+                // Compute the error.
+                let error = self.refresh_interval.as_micros() as i64
+                    - last_frame_timestamp.elapsed().as_micros() as i64;
+                let new_swap_delay = swap_delay.as_micros() as i64 + error;
+                Duration::from_micros(u64::try_from(new_swap_delay).unwrap_or_default())
+            },
+            // If we don't have any information about the last frame, just return
+            // previous target.
+            None => swap_delay,
+        };
+
+        self.swap_delay = Some(new_swap_delay);
+        self.frame_timestamp = Some(Instant::now());
+
+        new_swap_delay
     }
 }